requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
apscheduler==3.10.0
asyncio==3.4.3
beautifulsoup4==4.11.2
bing_image_downloader==1.1.2
bleach==6.0.0
bs4==0.0.1
cachetools==5.3.0
countryinfo==0.1.2
currencyconverter==0.17.5
dateparser==1.1.7
emoji==2.2.0
emoji-country-flag==1.3.2
feedparser==6.0.10
future==0.18.3
geopy==2.3.0
gpytranslate==1.5.1
hachoir==3.2.0
html2text==2020.1.16
jikanpy==4.3.2
lxml==4.9.2
markdown2==2.4.7
motor==3.1.1
opencv-python-headless==4.7.0.68
pillow==9.4.0
pretty_errors==1.2.25
psutil==5.9.4
psycopg2-binary==2.9.5
pydictionary==2.0.1
pykeyboard==0.1.5
pynewtonmath==1.0.1
pyrate-limiter==2.8.5
pyrogram==2.0.97
python-telegram-bot==13.11
regex==2022.10.31
requests==2.28.2
secureme==0.0.4
search_engine_parser==0.6.8
speedtest-cli==2.1.3
sqlalchemy==1.4.46
telegraph==2.2.0
telethon==1.27.0
tgcrypto==1.2.5
wikipedia==1.4.0
webcolors
gettext
pony
